+++ v2/types/factories/IERC20Custody.sol/IERC20Custody__factory.ts
@@ -115,8 +115,8 @@ const _abi = [
           },
           {
             name: "amount",
-            type: "uint64",
-            internalType: "uint64",
+            type: "uint256",
+            internalType: "uint256",
           },
           {
             name: "revertMessage",
@@ -296,8 +296,8 @@ const _abi = [
           },
           {
             name: "amount",
-            type: "uint64",
-            internalType: "uint64",
+            type: "uint256",
+            internalType: "uint256",
           },
           {
             name: "revertMessage",
