+++ v2/types/factories/ZetaConnectorBase__factory.ts
@@ -343,8 +343,8 @@ const _abi = [
           },
           {
             name: "amount",
-            type: "uint64",
-            internalType: "uint64",
+            type: "uint256",
+            internalType: "uint256",
           },
           {
             name: "revertMessage",
@@ -563,8 +563,8 @@ const _abi = [
           },
           {
             name: "amount",
-            type: "uint64",
-            internalType: "uint64",
+            type: "uint256",
+            internalType: "uint256",
           },
           {
             name: "revertMessage",
